+++ integrations/azure_ai_search/tests/test_bm25_retriever.py
@@ -47,7 +47,7 @@ def test_to_dict():
                     "api_key": {"type": "env_var", "env_vars": ["AZURE_AI_SEARCH_API_KEY"], "strict": False},
                     "index_name": "default",
                     "embedding_dimension": 768,
-                    "metadata_fields": None,
+                    "metadata_fields": {},
                     "vector_search_configuration": {
                         "profiles": [
                             {"name": "default-vector-config", "algorithm_configuration_name": "cosine-algorithm-config"}
@@ -149,8 +149,8 @@ def test_run_time_params():
 
 
 @pytest.mark.skipif(
-    not os.environ.get("AZURE_SEARCH_SERVICE_ENDPOINT", None) and not os.environ.get("AZURE_SEARCH_API_KEY", None),
-    reason="Missing AZURE_SEARCH_SERVICE_ENDPOINT or AZURE_SEARCH_API_KEY.",
+    not os.environ.get("AZURE_AI_SEARCH_ENDPOINT", None) and not os.environ.get("AZURE_AI_SEARCH_API_KEY", None),
+    reason="Missing AZURE_AI_SEARCH_ENDPOINT or AZURE_AI_SEARCH_API_KEY.",
 )
 @pytest.mark.integration
 class TestRetriever: