+++ integrations/github/src/haystack_integrations/tools/github/__init__.py
@@ -1,0 +1,16 @@
+# SPDX-FileCopyrightText: 2022-present deepset GmbH <info@deepset.ai>
+#
+# SPDX-License-Identifier: Apache-2.0
+from .file_editor_tool import GitHubFileEditorTool
+from .issue_commenter_tool import GitHubIssueCommenterTool
+from .issue_viewer_tool import GitHubIssueViewerTool
+from .pr_creator_tool import GitHubPRCreatorTool
+from .repo_viewer_tool import GitHubRepoViewerTool
+
+__all__ = [
+    "GitHubFileEditorTool",
+    "GitHubIssueCommenterTool",
+    "GitHubIssueViewerTool",
+    "GitHubPRCreatorTool",
+    "GitHubRepoViewerTool",
+]