+++ integrations/stackit/src/haystack_integrations/components/embedders/stackit/text_embedder.py
@@ -1,9 +1,9 @@
 # SPDX-FileCopyrightText: 2023-present deepset GmbH <info@deepset.ai>
 #
 # SPDX-License-Identifier: Apache-2.0
-from typing import Optional
+from typing import Any, Dict, Optional
 
-from haystack import component
+from haystack import component, default_to_dict
 from haystack.components.embedders import OpenAITextEmbedder
 from haystack.utils.auth import Secret
 
@@ -30,6 +30,10 @@ class STACKITTextEmbedder(OpenAITextEmbedder):
         api_base_url: Optional[str] = "https://api.openai-compat.model-serving.eu01.onstackit.cloud/v1",
         prefix: str = "",
         suffix: str = "",
+        *,
+        timeout: Optional[float] = None,
+        max_retries: Optional[int] = None,
+        http_client_kwargs: Optional[Dict[str, Any]] = None,
     ):
         """
         Creates a STACKITTextEmbedder component.
@@ -45,6 +49,15 @@ class STACKITTextEmbedder(OpenAITextEmbedder):
             A string to add to the beginning of each text.
         :param suffix:
             A string to add to the end of each text.
+        :param timeout:
+            Timeout for STACKIT client calls. If not set, it defaults to either the `OPENAI_TIMEOUT` environment
+            variable, or 30 seconds.
+        :param max_retries:
+            Maximum number of retries to contact STACKIT after an internal error.
+            If not set, it defaults to either the `OPENAI_MAX_RETRIES` environment variable, or set to 5.
+        :param http_client_kwargs:
+            A dictionary of keyword arguments to configure a custom `httpx.Client`or `httpx.AsyncClient`.
+            For more information, see the [HTTPX documentation](https://www.python-httpx.org/api/#client).
         """
         super(STACKITTextEmbedder, self).__init__(  # noqa: UP008
             api_key=api_key,
@@ -54,4 +67,28 @@ class STACKITTextEmbedder(OpenAITextEmbedder):
             organization=None,
             prefix=prefix,
             suffix=suffix,
+            timeout=timeout,
+            max_retries=max_retries,
+            http_client_kwargs=http_client_kwargs,
+        )
+        # We add these since they were only added in Haystack 2.14.0
+        self.timeout = timeout
+        self.max_retries = max_retries
+
+    def to_dict(self) -> Dict[str, Any]:
+        """
+        Serializes the component to a dictionary.
+        :returns:
+            Dictionary with serialized data.
+        """
+        return default_to_dict(
+            self,
+            api_key=self.api_key.to_dict(),
+            model=self.model,
+            api_base_url=self.api_base_url,
+            prefix=self.prefix,
+            suffix=self.suffix,
+            timeout=self.timeout,
+            max_retries=self.max_retries,
+            http_client_kwargs=self.http_client_kwargs,
         )