+++ integrations/cohere/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [integrations/cohere-v4.2.1] - 2025-05-27
+
+### 🌀 Miscellaneous
+
+- Add pins for Cohere (#1817)
+
+## [integrations/cohere-v4.2.0] - 2025-05-26
+
+### 🚀 Features
+
+- Add run_async for CohereChatGenerator (#1689)
+
+
 ## [integrations/cohere-v4.1.0] - 2025-05-05
 
 ### 🚀 Features